requirements.txt
Django==1.11.7
djangorestframework==3.7.1
psycopg2==2.7.3.2
pytz==2017.3
